% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-project.R
\docType{data}
\name{ProjectReporter}
\alias{ProjectReporter}
\title{Project reporter: send test results to IRkernel, for use with DataCamp Projects.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ProjectReporter
}
\description{
This reporter subclasses the ListReporter, to send detailed test information for
use with jupyter notebooks.
}
\keyword{datasets}
