% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter-project.R
\name{run_tests}
\alias{run_tests}
\title{Creates a project reporter and executes tests.}
\usage{
run_tests(test_expr)
}
\arguments{
\item{test_expr}{testthat code to execute}
}
\description{
Creates a project reporter and executes tests.
}
